ENH: move CTLN code into src
