+++ src/AlgebraicDynamics.jl
@@ -5,9 +5,11 @@ using Reexport
 include("uwd_dynam.jl")
 include("dwd_dynam.jl")
 include("cpg_dynam.jl")
+include("ThresholdLinear/ThresholdLinear.jl")
 
 @reexport using .DWDDynam
 @reexport using .UWDDynam
 @reexport using .CPortGraphDynam
+@reexport using .ThresholdLinear
 
 end # module
